refactor: rename convex/lib/github.ts to github-client.ts
The low-level GitHub API client + crypto/JWT/webhook-signature helpers shared
the filename 'github.ts' with the cRPC endpoint module (convex/functions/
github.ts), forcing readers to check the path every time. Rename the lib to
github-client.ts and update its 4 importers. No behavior change.
